LPS-146970 Separate the message bus destination creation into a MessagingConfigurator class to avoid double-registration problems caused by exceptions thrown in the activate method
