+++ modules/apps/portal-background-task/portal-background-task-service/src/main/java/com/liferay/portal/background/task/internal/BackgroundTaskManagerImpl.java
@@ -30,13 +30,11 @@ import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.lock.LockManager;
 import com.liferay.portal.kernel.messaging.Destination;
-import com.liferay.portal.kernel.messaging.DestinationConfiguration;
 import com.liferay.portal.kernel.messaging.DestinationFactory;
 import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.MessageBus;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.util.ClassUtil;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.OrderByComparator;
 
 import java.io.File;
@@ -45,17 +43,12 @@ import java.io.Serializable;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Dictionary;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -606,11 +599,8 @@ public class BackgroundTaskManagerImpl implements BackgroundTaskManager {
 
 	@Activate
 	protected void activate(BundleContext bundleContext) {
-		_bundleContext = bundleContext;
-
-		Destination backgroundTaskDestination = _registerDestination(
-			bundleContext, DestinationConfiguration.DESTINATION_TYPE_PARALLEL,
-			DestinationNames.BACKGROUND_TASK, 5, 10);
+		Destination backgroundTaskDestination = _messageBus.getDestination(
+			DestinationNames.BACKGROUND_TASK);
 
 		BackgroundTaskMessageListener backgroundTaskMessageListener =
 			new BackgroundTaskMessageListener(
@@ -620,9 +610,8 @@ public class BackgroundTaskManagerImpl implements BackgroundTaskManager {
 
 		backgroundTaskDestination.register(backgroundTaskMessageListener);
 
-		Destination backgroundTaskStatusDestination = _registerDestination(
-			bundleContext, DestinationConfiguration.DESTINATION_TYPE_SERIAL,
-			DestinationNames.BACKGROUND_TASK_STATUS, 1, 1);
+		Destination backgroundTaskStatusDestination =
+			_messageBus.getDestination(DestinationNames.BACKGROUND_TASK_STATUS);
 
 		BackgroundTaskQueuingMessageListener
 			backgroundTaskQueuingMessageListener =
@@ -645,51 +634,8 @@ public class BackgroundTaskManagerImpl implements BackgroundTaskManager {
 		}
 	}
 
-	@Deactivate
-	protected void deactivate() {
-		for (ServiceRegistration<Destination> serviceRegistration :
-				_serviceRegistrations) {
-
-			Destination destination = _bundleContext.getService(
-				serviceRegistration.getReference());
-
-			serviceRegistration.unregister();
-
-			destination.destroy();
-		}
-
-		_bundleContext = null;
-	}
-
 	@Reference(unbind = "-")
 	protected void setLockManager(LockManager lockManager) {
-	}
-
-	private Destination _registerDestination(
-		BundleContext bundleContext, String destinationType,
-		String destinationName, int workersCoreSize, int workersMaxSize) {
-
-		DestinationConfiguration destinationConfiguration =
-			new DestinationConfiguration(destinationType, destinationName);
-
-		destinationConfiguration.setWorkersCoreSize(workersCoreSize);
-		destinationConfiguration.setWorkersMaxSize(workersMaxSize);
-
-		Destination destination = _destinationFactory.createDestination(
-			destinationConfiguration);
-
-		Dictionary<String, Object> dictionary =
-			HashMapDictionaryBuilder.<String, Object>put(
-				"destination.name", destination.getName()
-			).build();
-
-		ServiceRegistration<Destination> serviceRegistration =
-			bundleContext.registerService(
-				Destination.class, destination, dictionary);
-
-		_serviceRegistrations.add(serviceRegistration);
-
-		return destination;
 	}
 
 	private List<BackgroundTask> _translate(
@@ -746,12 +692,14 @@ public class BackgroundTaskManagerImpl implements BackgroundTaskManager {
 	private BackgroundTaskLocalService _backgroundTaskLocalService;
 
 	@Reference
+	private BackgroundTaskMessagingConfigurator
+		_backgroundTaskMessagingConfigurator;
+
+	@Reference
 	private BackgroundTaskStatusRegistry _backgroundTaskStatusRegistry;
 
 	@Reference
 	private BackgroundTaskThreadLocalManager _backgroundTaskThreadLocalManager;
-
-	private volatile BundleContext _bundleContext;
 
 	@Reference
 	private ClusterMasterExecutor _clusterMasterExecutor;
@@ -764,8 +712,5 @@ public class BackgroundTaskManagerImpl implements BackgroundTaskManager {
 
 	@Reference
 	private MessageBus _messageBus;
-
-	private final Set<ServiceRegistration<Destination>> _serviceRegistrations =
-		new HashSet<>();
 
 }
